[checkpoint_engine] fold DeltaFlush into encode.py, drop stale wrapper.py
wrapper.py's module docstring still described the deleted pure-delta
generator wrapper; all that remained was the DeltaFlush container. Move
it next to DeltaParam/checksum in encode.py (one on-wire schema module)
and refresh encode.py's docstring, which still advertised the removed
'deltas' gap-encoding.
