+++ verl/checkpoint_engine/delta_sync/__init__.py
@@ -23,8 +23,7 @@ Design follows THUDM/slime's delta-sync implementation
 (``slime/backends/megatron_utils/update_weight/update_weight_from_distributed_delta.py``).
 """
 
-from .encode import DeltaParam, checksum
-from .wrapper import DeltaFlush
+from .encode import DeltaFlush, DeltaParam, checksum
 
 __all__ = [
     "DeltaFlush",
